+++ app/tools/text/__init__.py
@@ -5,5 +5,13 @@ from .ask_gemini import ask_gemini
 from .code_review import code_review
 from .brainstorm import brainstorm
 from .challenge import challenge
+from .conversations import list_conversations, delete_conversation
 
-__all__ = ["ask_gemini", "code_review", "brainstorm", "challenge"]
+__all__ = [
+    "ask_gemini",
+    "code_review",
+    "brainstorm",
+    "challenge",
+    "list_conversations",
+    "delete_conversation"
+]